WIP: Phase 2 - Static allocation with ASID and Pluck support
- Convert thread management to static arrays (MaxThreads=64)
- Change TID from int32 to int16 for ASID support
- TID now equals slot index (no separate allocator)
- Add StaticRingBuf with InUse markers for hole support
- Implement Pluck() for mid-queue removal
- Update ThreadBlockFutex/Sleep to pluck from ready queue
- Update launch.go to use threadList.Allocate()
- Add interrupt-driven disk I/O design documentation

+++ src/kmazarin/golang/ksyscall/clone.go
@@ -8,7 +8,8 @@ import (
 
 // CloneThread is provided by main package via go:linkname.
 // Creates a new thread with proper context for Go's clone wrapper.
-func CloneThread(stack, returnAddr, spsr, mp, gp, fn uint64) int32
+// Returns TID (int16 = slot index, used as ASID for VM).
+func CloneThread(stack, returnAddr, spsr, mp, gp, fn uint64) int16
 
 // GetSyscallELR is provided by main package via go:linkname.
 // Returns the ELR_EL1 (return address) for the current syscall.
